+++ src/amino.ts
@@ -1,7 +1,4 @@
-import {
-  AminoConverters,
-  createDefaultAminoConverters,
-} from '@cosmjs/stargate';
+import { createDefaultAminoConverters } from '@cosmjs/stargate';
 import { createWasmAminoConverters } from '@cosmjs/cosmwasm-stargate';
 import { cosmosAminoConverters } from '@neutron-org/neutronjs/cosmos/client';
 import * as neutronContractmanagerTxAmino from '@neutron-org/neutronjs/neutron/contractmanager/tx.amino';
